[BugFix] 좋아요 두번누르면 안되는 문제 해결
- 반대로 지정한 함수 조정
