+++ src/components/Atoms/Inputs/ChatTextarea/ChatTextarea.tsx
@@ -1,7 +1,8 @@
 import { Svg } from '@/components/Atoms';
+import { useEffect } from 'react';
 
 interface TChatTexarea {
-  forwardRef: React.Ref<HTMLTextAreaElement>;
+  forwardRef: React.MutableRefObject<HTMLTextAreaElement | null>;
   label: string;
   id: string;
   name: string;
@@ -10,12 +11,12 @@ interface TChatTexarea {
   required?: boolean;
   rows?: number;
   className?: string;
-  onClick: (e: React.MouseEvent<HTMLButtonElement>) => void;
+  onClick: (e: React.MouseEvent<HTMLButtonElement> | KeyboardEvent) => void;
 }
 
-function ChatTextarea({
+const ChatTextarea = ({
   label,
-  forwardRef = null,
+  forwardRef,
   id,
   name,
   placeholder,
@@ -24,13 +25,56 @@ function ChatTextarea({
   rows = 1,
   className,
   onClick,
-}: TChatTexarea) {
+}: TChatTexarea) => {
   const textareaStyle = {
     className:
       'flex min-h-[40px] flex-row gap-4 rounded-5xl border-[1px] border-bjgray-100 bg-bjgray-100 px-4 py-2 focus-within:border-bjgray-500 has-[:disabled]:bg-bjgray-200',
     textarea:
       'resize-none h-auto overflow-y-hidden whitespace-pre-wrap bg-bjgray-100 text-b-1-regular text-bjblack placeholder:text-b-1-regular placeholder:text-bjgray-500 focus:outline-none disabled:text-bjgray-500',
   };
+
+  // textarea 크기 조절
+  useEffect(() => {
+    const { current: textarea } = forwardRef;
+    if (textarea) {
+      const handleInput = () => {
+        textarea.style.height = 'auto'; // 높이 초기화
+        textarea.style.height = textarea.scrollHeight + 'px'; // 스크롤 높이에 맞춰 높이 설정
+      };
+
+      textarea.addEventListener('input', handleInput);
+      return () => textarea.removeEventListener('input', handleInput);
+    }
+  }, []);
+
+  // 보내기 버튼 키보드로 작동
+  useEffect(() => {
+    const { current: textarea } = forwardRef;
+
+    const handleKeydown = (e: KeyboardEvent) => {
+      const isPressedEnterKey = e.key === 'Enter';
+      const isPressedShiftKey = e.shiftKey;
+
+      if (isPressedShiftKey && isPressedEnterKey) {
+        // console.log('Shift + Enter 눌렀을 때 ');
+        textarea!.value += '\n';
+      }
+      if (isPressedEnterKey && !isPressedShiftKey) {
+        // console.log('Enter만 눌렀을 때');
+        e.preventDefault();
+        if (!textarea?.value) {
+          return;
+        }
+        onClick(e);
+      }
+    };
+
+    textarea?.addEventListener('keydown', handleKeydown);
+
+    return () => {
+      textarea?.removeEventListener('keydown', handleKeydown);
+    };
+  }, []);
 
   return (
     <div className={`${textareaStyle.className} ${className}`}>
@@ -47,12 +91,6 @@ function ChatTextarea({
           required={required}
           rows={rows}
           className={textareaStyle.textarea}
-          // defaultValue={defaultValue}
-          // disabled={disabled}
-          // readOnly={readOnly}
-          // minLength={minLength}
-          // maxLength={maxLength}
-          // {...rest}
         />
       </div>
       <button
@@ -65,6 +103,6 @@ function ChatTextarea({
       </button>
     </div>
   );
-}
+};
 
 export default ChatTextarea;
